chore(dashboard): move directories
