+++ dashboard/src/app/[tenantId]/components/CopyButton.tsx
